Enable use generics macro on impl, struct and trait
It's still needed to specify the list of generics (without bounds)
in the 'path' part of an impl block. This could be done automatically
if the structure is always like:

  impl<X: T1, Y: T2, Z: T3> Foo<X, Y, Z> { }

but there might be cases like:

  impl<X: T1 + T2> for Foo<Bar<X>> { }
or
  impl<'a, T> IntoIterator for &'a mut Option<T>

so the injection of generics to path of an impl cannot be generalized.

Closes #1